rename dir to hooks
