+++ packages/react/tsup.config.ts
@@ -2,8 +2,8 @@ import { defineConfig } from "tsup";
 
 export default defineConfig({
   entry: [
-    "src/react/credits.ts",
-    "src/react/buckets.ts",
+    "src/hooks/credits.ts",
+    "src/hooks/buckets.ts",
     "src/actions/credits.ts",
   ],
   format: ["cjs", "esm"],
